+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # MeasureMap Changelog
 
 ## vNext
+- Add ThreadId and Iteration to the Tracedetails
 
 ## v1.7.0
 - Set the duration that a Profilersession should run for